Extract TestExecutionEngine plugin event tests to partial class
